fltk: use size_t for index types, so gcc doesn't complain about signed and unsigned comparison

+++ fltk/EventTrack.cc
@@ -832,7 +832,7 @@ draw_trigger(bool draw_text, int x, double y, int w, const Event &event,
 */
 static int
 drawable_pixels(
-    int index, const Event *events, const vector<EventTrack::TextBox> &boxes)
+    size_t index, const Event *events, const vector<EventTrack::TextBox> &boxes)
 {
     const Event &event = events[index];
     if (!event.text || !*event.text)
@@ -864,7 +864,7 @@ drawable_pixels(
                 event_box.w += 2;
             }
             // Go forward to find all events starting here.
-            int prev = index + 1;
+            size_t prev = index + 1;
             while (prev < boxes.size() && events[prev].start == event.start)
                 prev++;
             for (--prev; prev >= 0; --prev) {
@@ -924,7 +924,7 @@ drawable_pixels(
                 event_box.w += 2;
             }
             // Rewind to find all events starting here.
-            int next = index - 1;
+            size_t next = index - 1;
             while (next >= 0 && events[next].start == event.start)
                 --next;
             for (++next; next < boxes.size(); ++next) {
